fix: escaping was broken
- innerText not supported by jsdom (it's nonstandard apparently)
- also, add docs on the main function
- also, the generate() function was supposed to be in script.js
