+++ package.json
@@ -17,6 +17,7 @@
   },
   "homepage": "https://github.com/not-dalia/gfm-toc#readme",
   "dependencies": {
+    "escape-html": "^1.0.3",
     "jsdom": "^19.0.0"
   },
   "devDependencies": {
